Code review changes - move params to API command and fix some ui layout
