+++ ui/index.html
@@ -99,19 +99,19 @@
                                     <p><translate key="message.select.a.zone"/></p>
                                     <div class="select-area lower-area">
                                         <select name="zoneid" class="zoneid required" required></select>
-                                        <label for="zoneid" class="description"><translate key="label.zone"/></label>
+                                        <label for="zoneid" class="desc"><translate key="label.zone"/></label>
                                     </div>
                                     <div class="select-area lower-area">
                                         <select name="podid" class="podid"></select>
-                                        <label for="podid" class="description"><translate key="label.pod"/></label>
+                                        <label for="podid" class="desc"><translate key="label.pod"/></label>
                                     </div>
                                     <div class="select-area lower-area">
                                         <select name="clusterid" class="clusterid"></select>
-                                        <label for="clusterid" class="description"><translate key="label.cluster"/></label>
+                                        <label for="clusterid" class="desc"><translate key="label.cluster"/></label>
                                     </div>
                                     <div class="select-area lower-area">
                                         <select name="hostid" class="hostid"></select>
-                                        <label for="hostid" class="description"><translate key="label.host"/></label>
+                                        <label for="hostid" class="desc"><translate key="label.host"/></label>
                                     </div>
                                 </div>
                                 <!-- Select template -->
